crypto: Fix package imports from the refactor
