+++ crypto/armor/armor_test.go
@@ -1,4 +1,4 @@
-package crypto
+package armor
 
 import (
 	"testing"
