+++ crypto/encoding/amino/encode_test.go
@@ -1,4 +1,4 @@
-package crypto
+package cryptoAmino
 
 import (
 	"os"
