+++ crypto/xsalsa20symmetric/symmetric.go
@@ -1,4 +1,4 @@
-package crypto
+package xsalsa20symmetric
 
 import (
 	"errors"
