+++ crypto/xsalsa20symmetric/symmetric_test.go
@@ -1,4 +1,4 @@
-package crypto
+package xsalsa20symmetric
 
 import (
 	"testing"
